+++ runtime/vm/service.h
@@ -13,6 +13,9 @@
 
 namespace dart {
 
+#define SERVICE_PROTOCOL_MAJOR_VERSION 3
+#define SERVICE_PROTOCOL_MINOR_VERSION 5
+
 class Array;
 class EmbedderServiceHandler;
 class Error;