+++ sdk/lib/developer/developer.dart
@@ -21,6 +21,7 @@ import 'dart:convert';
 part 'extension.dart';
 part 'profiler.dart';
 part 'timeline.dart';
+part 'service.dart';
 
 /// If [when] is true, stop the program as if a breakpoint were hit at the
 /// following statement.